Add PlatformNativeTokenProvider using the AuthenticationServices framework (#55)
* oauth2: add PlatformNativeTokenProvider

The platform provider uses the native AuthenticationServices framework
to handle the authenitication request. This means that the terminal
redirect will get sent back into the app instead of landing the user
on a page in the browser. This provider is macOS/iOS only.

* example: add macOS/iOS native demo

Add examples of the embedded browser and the AuthenticationServices
flows in an xcodeproject. The macOS flow isn't working at the moment.

* fixup: add more gauth setup details

* And move logic to its own file.

* example: add token caching